+++ crates/debugger/src/tui/builder.rs
@@ -1,10 +1,9 @@
 //! TUI debugger builder.
 
-use crate::{DebugNode, Debugger};
-use alloy_primitives::{Address, Bytes};
+use crate::{node::flatten_call_trace, DebugNode, Debugger};
+use alloy_primitives::Address;
 use foundry_common::{compile::ContractSources, evm::Breakpoints, get_contract_name};
 use foundry_evm_traces::{CallTraceArena, CallTraceDecoder, Traces};
-use revm_inspectors::tracing::types::TraceMemberOrder;
 use std::collections::HashMap;
 
 /// Debugger builder.
@@ -89,55 +88,5 @@ impl DebuggerBuilder {
     pub fn build(self) -> Debugger {
         let Self { debug_arena, identified_contracts, sources, breakpoints } = self;
         Debugger::new(debug_arena, identified_contracts, sources, breakpoints)
-    }
-}
-
-fn flatten_call_trace(arena: CallTraceArena, out: &mut Vec<DebugNode>) {
-    #[derive(Debug, Clone, Copy)]
-    struct PendingNode {
-        node_idx: usize,
-        steps_count: usize,
-    }
-
-    fn inner(arena: &CallTraceArena, node_idx: usize, out: &mut Vec<PendingNode>) {
-        let mut pending = PendingNode { node_idx, steps_count: 0 };
-        let node = &arena.nodes()[node_idx];
-        for order in node.ordering.iter() {
-            match order {
-                TraceMemberOrder::Call(idx) => {
-                    out.push(pending);
-                    pending.steps_count = 0;
-                    inner(arena, node.children[*idx], out);
-                }
-                TraceMemberOrder::Step(_) => {
-                    pending.steps_count += 1;
-                }
-                _ => {}
-            }
-        }
-        out.push(pending);
-    }
-    let mut nodes = Vec::new();
-    inner(&arena, 0, &mut nodes);
-
-    let mut arena_nodes = arena.into_nodes();
-
-    for pending in nodes {
-        let steps = {
-            let other_steps =
-                arena_nodes[pending.node_idx].trace.steps.split_off(pending.steps_count);
-            std::mem::replace(&mut arena_nodes[pending.node_idx].trace.steps, other_steps)
-        };
-
-        // Skip nodes with empty steps as there's nothing to display for them.
-        if steps.is_empty() {
-            continue
-        }
-
-        let call = &arena_nodes[pending.node_idx].trace;
-        let calldata = if call.kind.is_any_create() { Bytes::new() } else { call.data.clone() };
-        let node = DebugNode::new(call.address, call.kind, steps, calldata);
-
-        out.push(node);
     }
 }
